use TypeScript for jest config files
