+++ docs/repo-docs/reference/run.mdx
@@ -283,7 +283,7 @@ Specify glob of global filesystem dependencies to be hashed. Useful for `.env` a
 
 ```bash title="Terminal"
 turbo run build --global-deps=".env"
-turbo run build --global-deps=".env.*" --global-deps=".eslintrc" --global-deps="jest.config.js"
+turbo run build --global-deps=".env.*" --global-deps=".eslintrc" --global-deps="jest.config.ts"
 ```
 
 <Callout type="info">
