+++ examples/kitchen-sink/packages/jest-presets/node/jest-preset.ts
@@ -1,4 +1,6 @@
-module.exports = {
+import type { Config } from "jest";
+
+const config = {
   roots: ["<rootDir>"],
   transform: {
     "^.+\\.tsx?$": "ts-jest",
@@ -10,4 +12,6 @@ module.exports = {
     "<rootDir>/dist",
   ],
   preset: "ts-jest",
-};
+} as const satisfies Config;
+
+export default config;
