+++ examples/kitchen-sink/packages/jest-presets/package.json
@@ -4,10 +4,11 @@
   "private": true,
   "license": "MIT",
   "files": [
-    "browser/jest-preset.js",
-    "node/jest-preset.js"
+    "browser/jest-preset.ts",
+    "node/jest-preset.ts"
   ],
   "dependencies": {
+    "jest": "^29.7.0",
     "ts-jest": "^29.2.5"
   },
   "devDependencies": {
