+++ examples/with-docker/packages/jest-presets/package.json
@@ -4,9 +4,10 @@
   "private": true,
   "license": "MIT",
   "files": [
-    "node/jest-preset.js"
+    "node/jest-preset.ts"
   ],
   "dependencies": {
+    "jest": "^29.7.0",
     "ts-jest": "^29.2.5"
   }
 }
