--- examples/with-nestjs/apps/api/jest.config.js
@@ -1,4 +1,0 @@
-/** @type {import('jest').Config} */
-module.exports = {
-  ...require('@repo/jest-config/nest'),
-};
