+++ examples/with-nestjs/apps/api/jest.config.ts
@@ -1,0 +1,3 @@
+import { config } from '@repo/jest-config/nest';
+
+export default config;
