+++ examples/with-nestjs/apps/api/package.json
@@ -38,7 +38,7 @@
     "supertest": "^6.3.3",
     "ts-jest": "^29.2.5",
     "ts-loader": "^9.4.3",
-    "ts-node": "^10.9.1",
+    "ts-node": "10.9.2",
     "tsconfig-paths": "^4.2.0",
     "typescript": "5.4.5"
   }
