--- examples/with-nestjs/apps/web/jest.config.js
@@ -1,3 +1,0 @@
-const config = require('@repo/jest-config/next');
-
-module.exports = config();
