+++ examples/with-nestjs/apps/web/jest.config.ts
@@ -1,0 +1,3 @@
+import config from '@repo/jest-config/next';
+
+export default config;
