+++ examples/with-nestjs/packages/api/package.json
@@ -41,7 +41,7 @@
     "@repo/typescript-config": "workspace:*",
     "@types/node": "^20.3.1",
     "ts-loader": "^9.4.3",
-    "ts-node": "^10.9.1",
+    "ts-node": "10.9.2",
     "typescript": "5.4.5"
   }
 }
