+++ examples/with-nestjs/packages/jest-config/base.ts
@@ -1,0 +1,9 @@
+import type { Config } from 'jest';
+
+export const config = {
+  collectCoverage: true,
+  coverageDirectory: 'coverage',
+  coverageProvider: 'v8',
+  moduleFileExtensions: ['js', 'ts', 'json'],
+  testEnvironment: 'jsdom',
+} as const satisfies Config;
