+++ examples/with-nestjs/packages/jest-config/package.json
@@ -7,6 +7,7 @@
     "access": "public"
   },
   "devDependencies": {
+    "@jest/types": "^29.6.3",
     "jest": "^29.7.0",
     "next": "^14.1.1"
   }
