+++ packages/eslint-plugin-turbo/jest.config.ts
@@ -1,5 +1,6 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
+import type { Config } from "jest";
+
+const config = {
   roots: ["<rootDir>"],
   transform: {
     "^.+\\.tsx?$": "ts-jest",
@@ -11,4 +12,6 @@ module.exports = {
   preset: "ts-jest",
   verbose: process.env.RUNNER_DEBUG === "1",
   silent: process.env.RUNNER_DEBUG !== "1",
-};
+} as const satisfies Config;
+
+export default config;
