+++ packages/turbo-benchmark/package.json
@@ -40,6 +40,8 @@
     "@types/ndjson": "^2.0.2",
     "@types/node": "^18.17.4",
     "@types/node-fetch": "^2.6.6",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.5",
     "typescript": "5.4.5"
   }
 }
