+++ packages/turbo-codemod/jest.config.ts
@@ -1,5 +1,6 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
+import type { Config } from "jest";
+
+const config = {
   preset: "ts-jest/presets/js-with-ts",
   testEnvironment: "node",
   transformIgnorePatterns: ["node_modules/*", "packages/turbo-workspaces/*"],
@@ -27,4 +28,6 @@ module.exports = {
   },
   verbose: process.env.RUNNER_DEBUG === "1",
   silent: process.env.RUNNER_DEBUG !== "1",
-};
+} as const satisfies Config;
+
+export default config;
