+++ packages/turbo-gen/package.json
@@ -29,7 +29,7 @@
     "node-plop": "^0.26.3",
     "picocolors": "1.0.1",
     "proxy-agent": "^6.2.2",
-    "ts-node": "^10.9.1",
+    "ts-node": "^10.9.2",
     "update-check": "^1.5.4",
     "validate-npm-package-name": "^5.0.0"
   },
