+++ packages/turbo-utils/jest.config.ts
@@ -1,9 +1,12 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
+import type { Config } from "jest";
+
+const config = {
   preset: "ts-jest/presets/js-with-ts",
   testEnvironment: "node",
   modulePathIgnorePatterns: ["<rootDir>/node_modules", "<rootDir>/dist"],
   transformIgnorePatterns: ["/node_modules/(?!(ansi-regex)/)"],
   verbose: process.env.RUNNER_DEBUG === "1",
   silent: process.env.RUNNER_DEBUG !== "1",
-};
+} as const satisfies Config;
+
+export default config;
